requirements.txt
discord.py
pystyle
datetime